+++ app/src/main/java/com/example/appdevelopment/data/AuthViewModel.kt
@@ -1,0 +1,35 @@
+package com.example.appdevelopment.data
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.google.firebase.auth.FirebaseUser
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.launch
+import javax.inject.Inject
+
+class AuthViewModel @Inject constructor(
+    private val repository: AuthRepository
+) : ViewModel() {
+
+    private val _loginFlow = MutableStateFlow<Resource<FirebaseUser>?>(null)
+    val loginFlow: StateFlow<Resource<FirebaseUser>?> = _loginFlow
+
+    private val _signupFlow = MutableStateFlow<Resource<FirebaseUser>?>(null)
+    val signupFlow: StateFlow<Resource<FirebaseUser>?> = _signupFlow
+
+    val currentUser: FirebaseUser?
+    get() = repository.currentUser
+
+    init {
+        if(repository.currentUser != null){
+            _loginFlow.value = Resource.Success(repository.currentUser!!)
+        }
+    }
+
+    fun login(email: String, password: String) = viewModelScope.launch {
+        _loginFlow.value = Resource.Loading
+        val result = repository.login(email, password)
+        _loginFlow.value = result
+    }
+}
